smartju/smartju/urls.py
"""
URL configuration for smartju project.
"""
from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from django.http import JsonResponse
from django.views.decorators.http import require_http_methods
from rest_framework.routers import DefaultRouter
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)
from drf_yasg.views import get_schema_view
from drf_yasg import openapi

# Import ViewSets
from accounts.views import UserProfileViewSet
from lawsuits.views import LawsuitViewSet, LegalTemplateViewSet, FinancialClaimViewSet
from parties.views import PlaintiffViewSet, DefendantViewSet
from attachments.views import AttachmentViewSet
from responses.views import ResponseViewSet
from appeals.views import AppealViewSet
from hearings.views import HearingViewSet
from judgments.views import JudgmentViewSet
from audit.views import AuditLogViewSet
from courts.views import (
    GovernorateViewSet, DistrictViewSet, CourtTypeViewSet,
    CourtSpecializationViewSet, CourtViewSet
)
from payments.views import PaymentOrderViewSet
from laws.views import (
    LegalCategoryViewSet, LawViewSet, LawChapterViewSet,
    LawSectionViewSet, LawArticleViewSet, CaseLegalReferenceViewSet
)
from logs.views import UserSessionViewSet, SearchLogViewSet, AIChatLogViewSet

# Create router
router = DefaultRouter()
router.register(r'profiles', UserProfileViewSet, basename='profile')
router.register(r'lawsuits', LawsuitViewSet, basename='lawsuit')
router.register(r'legal-templates', LegalTemplateViewSet, basename='legal-template')
router.register(r'financial-claims', FinancialClaimViewSet, basename='financial-claim')
router.register(r'plaintiffs', PlaintiffViewSet, basename='plaintiff')
router.register(r'defendants', DefendantViewSet, basename='defendant')
router.register(r'attachments', AttachmentViewSet, basename='attachment')
router.register(r'responses', ResponseViewSet, basename='response')
router.register(r'appeals', AppealViewSet, basename='appeal')
router.register(r'hearings', HearingViewSet, basename='hearing')
router.register(r'judgments', JudgmentViewSet, basename='judgment')
router.register(r'audit-logs', AuditLogViewSet, basename='audit-log')
# Courts
router.register(r'governorates', GovernorateViewSet, basename='governorate')
router.register(r'districts', DistrictViewSet, basename='district')
router.register(r'court-types', CourtTypeViewSet, basename='court-type')
router.register(r'court-specializations', CourtSpecializationViewSet, basename='court-specialization')
router.register(r'courts', CourtViewSet, basename='court')
# Payments
router.register(r'payment-orders', PaymentOrderViewSet, basename='payment-order')
# Laws
router.register(r'legal-categories', LegalCategoryViewSet, basename='legal-category')
router.register(r'laws', LawViewSet, basename='law')
router.register(r'law-chapters', LawChapterViewSet, basename='law-chapter')
router.register(r'law-sections', LawSectionViewSet, basename='law-section')
router.register(r'law-articles', LawArticleViewSet, basename='law-article')
router.register(r'case-legal-references', CaseLegalReferenceViewSet, basename='case-legal-reference')
# Logs
router.register(r'user-sessions', UserSessionViewSet, basename='user-session')
router.register(r'search-logs', SearchLogViewSet, basename='search-log')
router.register(r'ai-chat-logs', AIChatLogViewSet, basename='ai-chat-log')

# Swagger schema view
schema_view = get_schema_view(
    openapi.Info(
        title="SmartJudi API",
        default_version='v1',
        description="""
        منصة قضائية - REST API
        
        هذه API توفر واجهة برمجية شاملة لإدارة النظام القضائي.
        
        ## التوثيق الكامل:
        
        ### المصادقة (Authentication):
        - استخدام JWT Token للمصادقة
        - الحصول على Token من `/api/token/`
        - استخدام Refresh Token من `/api/token/refresh/`
        
        ### الأدوار (Roles):
        - **judge**: قاضي - يمكنه إنشاء وتعديل الجلسات والأحكام
        - **lawyer**: محامي - يمكنه إنشاء وتعديل الدعاوى والأطراف
        - **notary**: كاتب عدل
        - **citizen**: مواطن - يمكنه فقط رؤية دعاويه
        - **admin**: مدير - صلاحيات كاملة
        
        ### Endpoints المتاحة:
        - `/api/profiles/` - ملفات المستخدمين
        - `/api/lawsuits/` - الدعاوى
        - `/api/plaintiffs/` - المدعون
        - `/api/defendants/` - المدعى عليهم
        - `/api/attachments/` - المرفقات
        - `/api/responses/` - الردود والمذكرات
        - `/api/appeals/` - الطعون
        - `/api/hearings/` - الجلسات
        - `/api/judgments/` - الأحكام
        - `/api/audit-logs/` - سجل الإجراءات (قراءة فقط)
        
        ### Pagination:
        جميع endpoints تدعم Pagination بحجم صفحة 20 عنصر.
        
        ### Filtering:
        جميع endpoints تدعم Filtering و Search و Ordering.
        """,
        terms_of_service="https://www.google.com/policies/terms/",
        contact=openapi.Contact(
            name="SmartJudi Support",
            email="contact@smartjudi.local"
        ),
        license=openapi.License(name="Proprietary License"),
    ),
    public=True,
    permission_classes=[],  # Allow access to schema without authentication
)

# Home page view
@require_http_methods(["GET"])
def home_view(request):
    """Home page that provides API information"""
    return JsonResponse({
        'message': 'مرحباً بك في منصة SmartJudi القضائية',
        'welcome': 'Welcome to SmartJudi Judicial Platform',
        'version': '1.0.0',
        'endpoints': {
            'api_documentation': '/swagger/',
            'api_documentation_redoc': '/redoc/',
            'api_base': '/api/',
            'admin_panel': '/admin/',
            'authentication': {
                'obtain_token': '/api/token/',
                'refresh_token': '/api/token/refresh/',
            },
            'resources': {
                'profiles': '/api/profiles/',
                'lawsuits': '/api/lawsuits/',
                'plaintiffs': '/api/plaintiffs/',
                'defendants': '/api/defendants/',
                'attachments': '/api/attachments/',
                'responses': '/api/responses/',
                'appeals': '/api/appeals/',
                'hearings': '/api/hearings/',
                'judgments': '/api/judgments/',
                'audit_logs': '/api/audit-logs/',
                'governorates': '/api/governorates/',
                'districts': '/api/districts/',
                'court_types': '/api/court-types/',
                'court_specializations': '/api/court-specializations/',
                'courts': '/api/courts/',
                'payment_orders': '/api/payment-orders/',
                'legal_categories': '/api/legal-categories/',
                'laws': '/api/laws/',
                'law_chapters': '/api/law-chapters/',
                'law_sections': '/api/law-sections/',
                'law_articles': '/api/law-articles/',
                'case_legal_references': '/api/case-legal-references/',
                'user_sessions': '/api/user-sessions/',
                'search_logs': '/api/search-logs/',
                'ai_chat_logs': '/api/ai-chat-logs/',
            }
        },
        'documentation': 'Visit /swagger/ for interactive API documentation',
    }, json_dumps_params={'ensure_ascii': False, 'indent': 2})


urlpatterns = [
    # Home page
    path('', home_view, name='home'),
    
    # Admin
    path('admin/', admin.site.urls),
    
    # API Routes
    path('api/', include(router.urls)),
    
    # JWT Authentication
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    
    # User Registration
    path('api/register/', include('accounts.urls')),
    
    # Swagger/OpenAPI Documentation
    path('swagger/', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
    path('redoc/', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
    path('swagger.json', schema_view.without_ui(cache_timeout=0), name='schema-json'),
]

# Serve media files in development
if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
    urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
