requirements.txt
Django>=6.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7
django-filter>=23.0
django-cors-headers>=4.3.0

